requirement.txt
absl-py==0.8.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
entrypoints==0.3
flake8==3.7.9
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mccabe==0.6.1
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pyflakes==2.1.1
pyserial==3.4
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.3.2
six==1.13.0
tb-nightly==2.1.0a20191119
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
tf-estimator-nightly==2.0.0.dev2019111909
tf-nightly==2.1.0.dev20191119
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2